refactor(content): retire the per type field admin
