+++ frontend/src/content/TypesScreen.tsx
@@ -7,7 +7,6 @@ import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
 import { useState } from 'react'
 import type { ReactNode } from 'react'
 
-import { FieldsDialog } from './FieldsDialog'
 import { typesQueryKey } from './nav'
 import { createType, deleteType, listTypes, updateType } from './types'
 import type { ContentType, TypeEdit } from './types'
@@ -130,7 +129,6 @@ function TypesBody(props: Reporter & { types: ContentType[]; loading: boolean; f
 							key={registered.key}
 							registered={registered}
 							holder={props.types.find((listed) => listed.isDefault)}
-							types={props.types}
 							onDone={props.onDone}
 							onRefused={props.onRefused}
 						/>
@@ -147,7 +145,7 @@ function TypesBody(props: Reporter & { types: ContentType[]; loading: boolean; f
  * @returns The row element.
  */
 function TypeRow(
-	props: Reporter & { registered: ContentType; holder?: ContentType; types: ContentType[] },
+	props: Reporter & { registered: ContentType; holder?: ContentType },
 ): ReactNode {
 	const { registered } = props
 	const edit = useMutation({
@@ -179,12 +177,6 @@ function TypeRow(
 			<td>
 				<Stack direction="row" gap="xs">
 					<ChangeAddress registered={registered} onMove={(word) => edit.mutate({ routeWord: word })} />
-					<FieldsDialog
-						registered={registered}
-						types={props.types}
-						onDone={props.onDone}
-						onRefused={props.onRefused}
-					/>
 					{!registered.isDefault && (
 						<HandOverRoot
 							registered={registered}
