+++ frontend/src/content/types.ts
@@ -250,118 +250,3 @@ export interface NewField {
 	required?: boolean
 }
 
-/**
- * Returns the fields a content type declares.
- * @param typeKey - The type whose fields to read.
- * @returns The declared fields.
- */
-export async function listFields(typeKey: string): Promise<ContentField[]> {
-	const response = await fetch(`/api/types/${typeKey}/fields`)
-	if (!response.ok) {
-		await refuse(response)
-	}
-	const listed = z.object({ items: z.array(fieldSchema) }).parse(await response.json())
-	return listed.items.map(toField)
-}
-
-/**
- * Declares a field on a content type.
- * @param typeKey - The type declaring the field.
- * @param asked - The field to declare.
- * @returns The stored field.
- */
-export async function createField(typeKey: string, asked: NewField): Promise<ContentField> {
-	const response = await fetch(`/api/types/${typeKey}/fields`, {
-		method: 'POST',
-		headers: { 'content-type': 'application/json' },
-		body: JSON.stringify({
-			key: asked.key,
-			label: asked.label,
-			kind: asked.kind,
-			relates_to: asked.relatesTo,
-			many: asked.many,
-			required: asked.required,
-		}),
-	})
-	if (!response.ok) {
-		await refuse(response)
-	}
-	return toField(fieldSchema.parse(await response.json()))
-}
-
-/**
- * Carries a new label for a declared field.
- * @param typeKey - The type declaring the field.
- * @param key - The field to relabel.
- * @param label - The label to carry.
- * @returns The stored field.
- */
-export async function renameField(
-	typeKey: string,
-	key: string,
-	label: string,
-): Promise<ContentField> {
-	const response = await fetch(`/api/types/${typeKey}/fields/${key}`, {
-		method: 'PATCH',
-		headers: { 'content-type': 'application/json' },
-		body: JSON.stringify({ label }),
-	})
-	if (!response.ok) {
-		await refuse(response)
-	}
-	return toField(fieldSchema.parse(await response.json()))
-}
-
-/**
- * Stores whether a field must be filled before its item publishes.
- * @param typeKey - The type declaring the field.
- * @param key - The field to change.
- * @param required - Whether the field gates publishing.
- * @returns The stored field.
- */
-export async function setFieldRequired(
-	typeKey: string,
-	key: string,
-	required: boolean,
-): Promise<ContentField> {
-	const response = await fetch(`/api/types/${typeKey}/fields/${key}`, {
-		method: 'PATCH',
-		headers: { 'content-type': 'application/json' },
-		body: JSON.stringify({ required }),
-	})
-	if (!response.ok) {
-		await refuse(response)
-	}
-	return toField(fieldSchema.parse(await response.json()))
-}
-
-/**
- * Stores the declaration order of a type's fields.
- * @param typeKey - The type whose fields are reordered.
- * @param keys - Every declared field key in the order to store.
- * @returns The reordered fields.
- */
-export async function reorderFields(typeKey: string, keys: string[]): Promise<ContentField[]> {
-	const response = await fetch(`/api/types/${typeKey}/fields/order`, {
-		method: 'PUT',
-		headers: { 'content-type': 'application/json' },
-		body: JSON.stringify({ order: keys }),
-	})
-	if (!response.ok) {
-		await refuse(response)
-	}
-	const listed = z.object({ items: z.array(fieldSchema) }).parse(await response.json())
-	return listed.items.map(toField)
-}
-
-/**
- * Removes a declared field and every value it held.
- * @param typeKey - The type declaring the field.
- * @param key - The field to remove.
- */
-export async function deleteField(typeKey: string, key: string): Promise<void> {
-	const response = await fetch(`/api/types/${typeKey}/fields/${key}`, { method: 'DELETE' })
-	if (!response.ok) {
-		await refuse(response)
-	}
-}
